+++ Objects/KaiaStructures/Exploration/Properties/Events/NotebookEvent.cs
@@ -10,7 +10,7 @@ namespace Kaia.Bot.Objects.KaiaStructures.Exploration.Properties.Events
 {
     public class NotebookEvent : KaiaLocationEvent
     {
-        public NotebookEvent(int Weight) : base(ExplorationStrings.NotebookEvent.Message, 
+        public NotebookEvent(double Weight) : base(ExplorationStrings.NotebookEvent.Message, 
             Weight, 
             new(0, new Notebook()))
         {
